refactor(settings): move mobile CountrySettingsTile into feature_iptv
Same module-ownership fix as PlaybackSettingsScreen/TvPlaybackSection/
TvSourceManagementSection: the country-filter tile is a pure IPTV
concern (channelFiltersProvider, iptvChannelsProvider,
channelFilterDimensions — all feature_iptv state), so it moves out of
app/lib/features/settings into feature_iptv as a public
CountrySettingsTile widget.

Completes the settings module-ownership pass started in #1101/#1103.
