+++ app/lib/features/settings/presentation/screens/settings_hub_screen.dart
@@ -1,11 +1,8 @@
-import 'dart:async';
-
 import 'package:core_product_shell/core_product_shell.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:core_ui/core_ui.dart';
 import 'package:feature_iptv/feature_iptv.dart';
-import 'package:feature_iptv/presentation/tv_ux/sections/filter_dialogs.dart';
 import '../../../../core/providers/app_theme_provider.dart';
 import 'audio_settings_screen.dart';
 
@@ -125,7 +122,7 @@ class SettingsHubScreen extends ConsumerWidget {
               },
             ),
 
-            const _CountrySettingsTile(),
+            const CountrySettingsTile(),
 
             ListTile(
               leading: Icon(
@@ -165,64 +162,3 @@ class SettingsHubScreen extends ConsumerWidget {
   }
 }
 
-class _CountrySettingsTile extends ConsumerWidget {
-  const _CountrySettingsTile();
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final filters = ref.watch(channelFiltersProvider);
-    final channelsAsync = ref.watch(iptvChannelsProvider);
-    final channels = channelsAsync.value ?? const <IPTVChannel>[];
-    final dimensions = channelFilterDimensions(
-      channels: channels,
-      metadataByChannelId: const {},
-    );
-    final canPickCountry =
-        dimensions.countries.isNotEmpty || filters.country != null;
-
-    final country = _section(IptvSettingsSectionId.country);
-    return ListTile(
-      leading: Icon(country.iconFor(ShellId.mobile)),
-      title: Text(country.labelFor(ShellId.mobile)),
-      subtitle: Text(
-        filters.country != null
-            ? countryDisplayLabel(filters.country)
-            : channelsAsync.isLoading
-            ? 'Loading countries…'
-            : dimensions.countries.isEmpty
-            ? 'Load channels first to choose a country'
-            : 'Choose your default channel country',
-      ),
-      trailing: const Icon(Icons.arrow_forward_ios, size: 16),
-      enabled: canPickCountry,
-      onTap: !canPickCountry
-          ? null
-          : () => _showCountryPicker(context, ref, dimensions, filters.country),
-    );
-  }
-
-  Future<void> _showCountryPicker(
-    BuildContext context,
-    WidgetRef ref,
-    ChannelFilterDimensions dimensions,
-    String? selectedCountry,
-  ) {
-    final filters = ref.read(channelFiltersProvider.notifier);
-    final countryPrompt = ref.read(channelCountryPromptProvider.notifier);
-    return showFilterOptionDialog(
-      context: context,
-      title: 'Country',
-      options: dimensions.countries.toList(growable: false),
-      selectedValue: selectedCountry,
-      onSelected: (country) {
-        filters.setCountry(country);
-        unawaited(countryPrompt.markCompleted());
-      },
-      onClear: () {
-        filters.setCountry(null);
-        unawaited(countryPrompt.markCompleted());
-      },
-      optionLabel: countryDisplayLabel,
-    );
-  }
-}
