+++ docs/wiki/3.-Navigating-Airo.md
@@ -98,9 +98,10 @@ Audio settings are mobile-only, and Accessibility plus a combined Sources
 section are TV-only — but any shared section stays in sync automatically
 across both screens.
 
-Section ownership follows the modular-app model: Playback and Source
-Management are IPTV-module concerns and live in the `feature_iptv` package
-(mobile's `PlaybackSettingsScreen` and TV's `TvPlaybackSection` /
+Section ownership follows the modular-app model: Playback, Source
+Management, and Country are IPTV-module concerns and live in the
+`feature_iptv` package (mobile's `PlaybackSettingsScreen` /
+`CountrySettingsTile` and TV's `TvPlaybackSection` /
 `TvSourceManagementSection`); Theme and Audio stay app-level "Airo profile"
 settings, since both are genuinely cross-app (Theme is shared with the whole
 app's chrome, Audio shares `audio_context_provider` with the Music player). A
